基于spring事件机制更新user-role关系

+++ easyblog-dao/src/main/java/top/easyblog/dao/auto/model/UserRoleRelationship.java
@@ -2,14 +2,12 @@ package top.easyblog.dao.auto.model;
 
 import java.util.Date;
 import lombok.Data;
-import lombok.EqualsAndHashCode;
-import lombok.ToString;
 
 @Data
 public class UserRoleRelationship {
-    private Integer userId;
+    private Long userId;
 
-    private Integer roleId;
+    private Long roleId;
 
     private Boolean enabled;
